+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Add `backup` subcommand to create a backup of your current configuration
+
 ## [1.5.0] - 2021-03-17
 ### Added
 - Add `use` subcommand as convenient shorthand for setting default `gcloud` project. With support of interactive or autocompletion `-project` mode.